requirements.txt
cachelib==0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Session==0.3.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
requests==2.23.0
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==1.0.1
gunicorn==20.0.4